gen1 and gen2 are sending their regards

+++ include/battle.h
@@ -819,7 +819,7 @@ struct BattleStruct
     s32 battlerExpReward;
     u16 prevTurnSpecies[MAX_BATTLERS_COUNT]; // Stores species the AI has in play at start of turn
     s16 moveDamage[MAX_BATTLERS_COUNT];
-    s8 critChance[MAX_BATTLERS_COUNT];
+    s16 critChance[MAX_BATTLERS_COUNT];
     u16 moveResultFlags[MAX_BATTLERS_COUNT];
     u8 missStringId[MAX_BATTLERS_COUNT];
     u8 noResultString[MAX_BATTLERS_COUNT];
